He creado el detalle de Alquiler y he modificado el atributo de alquiler que estaba en int tenia que ser decimal (#104)

+++ src/AppForSEII2526.API/Models/Alquiler.cs
@@ -32,7 +32,7 @@
         public int Periodo { get; set; }
 
         [Range(0, int.MaxValue, ErrorMessage = "El precio total no puede ser negativo")]
-        public int PrecioTotal { get; set; }
+        public decimal PrecioTotal { get; set; }
 
         public TiposMetodosPago MetodoPago { get; set; }
